chore: move /e2e to /tests/e2e
