+++ tests/e2e/Pages/AggregateComponenet.ts
