+++ tests/e2e/Pages/DisasterTypeComponent.ts
