+++ tests/e2e/tests/AggregatesSection/AllAggregatesComponentsPresentNotrigger.spec.ts
@@ -1,10 +1,9 @@
 import { test } from '@playwright/test';
+import AggregatesComponent from 'Pages/AggregateComponenet';
 import DashboardPage from 'Pages/DashboardPage';
-import MapComponent from 'Pages/MapComponent';
-import UserStateComponent from 'Pages/UserStateComponent';
 import { NoTriggerDataSet } from 'testData/testData.enum';
 
-import { FloodsScenario } from '../../../services/API-service/src/scripts/enum/mock-scenario.enum';
+import { FloodsScenario } from '../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
 import {
   getAccessToken,
   mockFloods,
@@ -17,9 +16,9 @@ let accessToken: string;
 test.beforeEach(async ({ page }) => {
   // Login
   const loginPage = new LoginPage(page);
+
   accessToken = await getAccessToken();
   await resetDB(accessToken);
-
   // We should maybe create one mock for all different disaster types for now we can just use floods
   await mockFloods(
     FloodsScenario.NoTrigger,
@@ -34,22 +33,15 @@ test.beforeEach(async ({ page }) => {
   );
 });
 
-test('[30538] All Map elements are present in no-trigger mode', async ({
+test('[30588] All Aggregate elements are present in no-trigger mode', async ({
   page,
 }) => {
   const dashboard = new DashboardPage(page);
-  const userState = new UserStateComponent(page);
-  const map = new MapComponent(page);
+  const aggregates = new AggregatesComponent(page);
 
   // Navigate to disaster type the data was mocked for
   await dashboard.navigateToFloodDisasterType();
   // Assertions
-  await userState.headerComponentIsVisible({
-    countryName: NoTriggerDataSet.CountryName,
-  });
-  await map.mapComponentIsVisible();
-  await map.breadCrumbViewIsVisible({ nationalView: true });
-  await map.isLegendOpen({ legendOpen: true });
-  await map.isLayerMenuOpen({ layerMenuOpen: false });
-  await map.assertAdminBoundariesVisible();
+  await aggregates.aggregateComponentIsVisible();
+  await aggregates.aggregatesAlementsDisplayedInNoTrigger();
 });
