+++ tests/e2e/tests/DisasterType/AllDisasterTypeComponentsPresentNotrigger.spec.ts
@@ -1,10 +1,10 @@
 import { test } from '@playwright/test';
-import ChatComponent from 'Pages/ChatComponent';
 import DashboardPage from 'Pages/DashboardPage';
+import DisasterTypeComponent from 'Pages/DisasterTypeComponent';
 import UserStateComponent from 'Pages/UserStateComponent';
 import { NoTriggerDataSet } from 'testData/testData.enum';
 
-import { FloodsScenario } from '../../../services/API-service/src/scripts/enum/mock-scenario.enum';
+import { FloodsScenario } from '../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
 import {
   getAccessToken,
   mockFloods,
@@ -34,12 +34,12 @@ test.beforeEach(async ({ page }) => {
   );
 });
 
-test('[30587] All Chat section elements are present in no-trigger mode', async ({
+test('[30586] All Disaster Type elements are present in no-trigger mode', async ({
   page,
 }) => {
   const dashboard = new DashboardPage(page);
   const userState = new UserStateComponent(page);
-  const chat = new ChatComponent(page);
+  const disasterType = new DisasterTypeComponent(page);
 
   // Navigate to disaster type the data was mocked for
   await dashboard.navigateToFloodDisasterType();
@@ -47,9 +47,7 @@ test('[30587] All Chat section elements are present in no-trigger mode', async (
   await userState.headerComponentIsVisible({
     countryName: NoTriggerDataSet.CountryName,
   });
-  await chat.chatColumnIsVisibleForNoTriggerState({
-    firstName: NoTriggerDataSet.firstName,
-    lastName: NoTriggerDataSet.lastName,
-  });
-  await chat.allChatButtonsArePresent();
+  await disasterType.topBarComponentIsVisible();
+  await disasterType.allDisasterTypeElementsArePresent();
+  await disasterType.timlineElementsAreVisible();
 });
